refactored code
Pulled out supabase requests from children componenets (last seen & status) and placed inside parent component friend card
